New: Finalizando crud de shipping instruction, alterações na migration e rotas na API

+++ backend/api/nascomex-manager-api/database/migrations/2024_06_12_224402_create_shopping_instruction_table.php
@@ -36,7 +36,7 @@ class CreateShoppingInstructionTable extends Migration
             $table->string('installment_loan', 255)->nullable();
             $table->unsignedBigInteger('banks_id');
             $table->string('enterprise_name', 255)->nullable();
-            $table->integer('enterprise_cnpj')->nullable();
+            $table->unsignedBigInteger('enterprise_cnpj')->nullable();
             $table->timestamps();
 
             $table->foreign('imports_id')->references('id')->on('imports')->onDelete('cascade');
